feat(functions): Added hard coded functions (#3)
* feat(functions): Added hard coded functions to get genres and artists

Wrote functions to return artist and genre data from the back end. In the future these will connect to the database, but this is a good stop gap so that front end development is not impeded.

* fix(functions): Allow all through CORS

Makes it work for the demo, but is massively insecure. The current data does not necessitate data security, but clean this up in the future.